misc fixes, fixture datetime naive warnings, rename module for consistency
