+++ server/apps/trak/serializers/contact_ser.py
@@ -1,7 +1,8 @@
 from rest_framework import serializers
 
 from apps.trak.models import Contact
-from apps.trak.serializers.trak_ser import TrakBaseSerializer
+
+from .base_ser import TrakBaseSerializer
 
 
 class EpaPhoneSerializer(TrakBaseSerializer):
